Add APC support
opensuse is stuck on PHP 5.4, so we can't use the PHP 5.5 built
in opcode cache.  Until we use PHP 5.5 (or base all of this on
a different system), we can use APC directly here.

As of this commit, this entire image generates a _clean_ phabricator
install with no "unresolved setup issues" except for the base
case of "enable people to create accounts."