Fix 'Configuration' links in README for Docker Hub (#41)
The README.md file for this project automatically becomes a [web page on Docker Hub](https://hub.docker.com/r/hachque/phabricator/) after a build.
 
The links for "Basic Configuration" and "Advanced Configuration" in README.md do not work properly on this generated page.  For example it points to this URL, which 404s:  https://hub.docker.com/r/hachque/phabricator/ADVANCED-CONFIG.md/

This changes the links to point to their GitHub blobs on master.  The other Markdown files don't require this because they do not appear on Docker Hub.